Refactor chacha20poly1305 module
