+++ Cargo.toml
@@ -19,8 +19,6 @@ bitcoin_hashes = { version = "0.13.0", default-features = false }
 
 [dev-dependencies]
 hex = "0.4.3"
-chacha20poly1305 = "0.10.1"
-chacha20 = "0.9.1"
 
 [lib]
 name = "bip324"
